Добавил в бд 5 уровней сложности, добавил саму сущность уровня сложности,  обновил бдвьюмодель и дао, так чтобы получить список уровней сложности, в навигации реализовал передачу индекса уровня сложности можду экранами сложности и упражнений

+++ app/src/main/java/com/example/clavtrain/ui/user/UserTrainingScreen.kt
@@ -199,7 +199,7 @@ fun UserTrainingScreen(
 private fun mainAlg(
     simbolsAmount: Int,
     maxMistakes: Int,
-    maxPressTime: Int
+    maxPressTime: Long
 ): String {
     var text = "";
     repeat (simbolsAmount) {
